Revert "Change stateful "actions" API to a faster/lighter "dispatch" strategy"
Although this improves the time spent in the xvdom framework, it's shifted to
the application code.

Most of the time, a components actions/reducer will be called from a view event.
Binding this event usually takes the form like this...

```
<button onclick={()=>dispatch(doStuff)}>do it</button>
```

... we've allocated function.  The old approach didn't have this.

Also, because we are passing a new function every render, it will cause a
rerender for the component you are passing this function to (equality check).

This reverts commit 5677478fb8a06ffb9efaef3f3c89f2eeea978c57.