Update createDynamic to update parentNode.
This small change in createDynamic's responsibilities has 2 benefits:

1) This allows the compiler to generate smaller code.

BEFORE:

```
_n.appendChild(inst.b = _xvdomCreateDynamic(true, _n, inst.a));
```

AFTER:

```
inst.b = _xvdomCreateDynamic(true, _n, inst.a);
```

2) Removes the need to create and hold a reference to a document fragment when rendering arrays.

Previously rendering array, used the document fragment as the context node.
The document fragment was a poor context node as it really was just a container
to hold the true context node (parent node or marker node), which was stored
as a custom 'xvdomContext' property.